etl_pipeline/tweet_collector/requirements.txt
tweepy
python-dotenv
pymongo
